pyproject.toml
[project]
name = "simplegif"
version = "0.1.0"
description = ""
authors = [{ name = "jefferson-norberto2", email = "jmn@cin.ufpe.br" }]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = "3.12.*"
dependencies = [
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "Pillow (>=10.4.0,<11.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
